Dev / Atividade Tarefa 3 Criando meu projeto spring

+++ organicos/src/main/java/com/ecommerce/organicos/repository/CategoriaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ecommerce.organicos.model.Categoria;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer>{
+public interface CategoriaRepository extends JpaRepository<Categoria, Long>{
 
 }
